pyproject.toml
[tool.poetry]
name = "kclassify"
version = "1.0.0"
description = "Tensorflow-Keras image classifier for transfer-learning training either locally or in AWS SageMaker."
authors = ["Danelle Cline <dcline@mbari.org>"]
readme = "README.md"
license = "GPL3"
homepage = "http://docs.mbari.org/kclassify"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
matplotlib = "3.3.4"
pillow = "8.3.2"
boto3 = "1.18.49"
scipy = "1.7.1"
scikit-image = "0.19.3"
scikit-learn  = "1.2.1"
tensorflow-gpu  = "2.6.0"
tensorflow-addons  = "0.16.1"
tensorflow-model-optimization = "0.7.3"
pandas = "1.5.3"
progressbar2 = "4.2.0"
wandb = "0.12.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "src/__init__.py:__version__",
    "pyproject.toml:version"
]

branch = "master"
upload_to_pypi = false
upload_to_release = true
build_command = "poetry build"

